requirement_pip.txt
pyyaml
pandas
netCDF4
scipy
matplotlib
pyecharts
seaborn
cdo
